+++ src/main/model/Carbohydrates.java
@@ -1,22 +1,21 @@
 package model;
 
 /** 
- * A basic class representing a quantity of Carbohydrates
- * 
+ * This class represents carbohydrates(macro-nutrient) present inside a food item
  * **/
 public class Carbohydrates {
 
     private double value;      //representing the value of carbohydrates
 
     //EFFECT: creates Carbohydrate object with its value set up as provided value
-    public Carbohydrates(int value) {
+    public Carbohydrates(double value) {
         this.value = value;
     }
 
     //REQUIRES: value >= 0
     //MODIFIES: This
     //EFFECT: set value field to given value 
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
